extract PaginationProps to root types.js
