+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Pagination/Pagination.js
@@ -1,20 +1,12 @@
 // @flow
 import React from 'react';
-import type {Element} from 'react';
 import Icon from '../Icon';
 import Loader from '../Loader';
+import type {PaginationProps} from '../../types.js';
 import {translate} from '../../utils/Translator';
 import paginationStyles from './pagination.scss';
 
-type Props = {
-    children: Element<*>,
-    current: ?number,
-    loading: boolean,
-    onChange: (page: number) => void,
-    total: ?number,
-};
-
-export default class Pagination extends React.PureComponent<Props> {
+export default class Pagination extends React.PureComponent<PaginationProps> {
     static defaultProps = {
         loading: false,
     };
