+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/Datagrid.js
@@ -2,7 +2,6 @@
 import {observer} from 'mobx-react';
 import {observable, action, computed} from 'mobx';
 import React from 'react';
-import type {ComponentType} from 'react';
 import equal from 'fast-deep-equal';
 import InfiniteScroller from '../../components/InfiniteScroller';
 import Pagination from '../../components/Pagination';
@@ -10,7 +9,6 @@ import DatagridStore from './stores/DatagridStore';
 import datagridAdapterRegistry from './registries/DatagridAdapterRegistry';
 import AbstractAdapter from './adapters/AbstractAdapter';
 import AdapterSwitch from './AdapterSwitch';
-import type {PaginationProps} from './types';
 
 type Props = {
     onItemClick?: (itemId: string | number) => void,
@@ -87,7 +85,7 @@ export default class Datagrid extends React.PureComponent<Props> {
         const page = store.getPage();
         const pageCount = store.pageCount;
         const Adapter = this.currentAdapter;
-        const PaginationAdapter: ComponentType<PaginationProps> = Adapter.getLoadingStrategy() === 'infiniteScroll'
+        const PaginationAdapter = Adapter.getLoadingStrategy() === 'infiniteScroll'
             ? InfiniteScroller
             : Pagination;
