+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/types.js
@@ -1,6 +1,5 @@
 // @flow
 import type {IObservableValue} from 'mobx'; // eslint-disable-line import/named
-import type {Element} from 'react';
 
 export type DataItem = {
     id: string | number,
@@ -17,14 +16,6 @@ export type DatagridAdapterProps = {
     onItemClick?: (itemId: string | number) => void,
     onItemSelectionChange?: (rowId: string | number, selected?: boolean) => void,
     onAllSelectionChange?: (selected?: boolean) => void,
-};
-
-export type PaginationProps = {
-    children: Element<*>,
-    current: ?number,
-    loading: boolean,
-    onChange: (page: number) => void,
-    total: ?number,
 };
 
 export type ObservableOptions = {
