+++ src/Sulu/Bundle/AdminBundle/Resources/js/types.js
@@ -1,0 +1,10 @@
+// @flow
+import type {Element} from 'react';
+
+export type PaginationProps = {
+    children: Element<*>,
+    current: ?number,
+    loading: boolean,
+    onChange: (page: number) => void,
+    total: ?number,
+};
